pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoreadme"
version = "0.1.0"
description = "Generates professional README files from your codebase using AI"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "autoreadme", email = "hello@autoreadme.dev" }]
keywords = ["readme", "documentation", "ai", "llm", "developer-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pathspec>=0.12.1",
]

[project.optional-dependencies]
ollama = ["ollama>=0.3"]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-generativeai>=0.7"]
all = [
    "ollama>=0.3",
    "openai>=1.0",
    "anthropic>=0.30",
    "google-generativeai>=0.7",
]

[project.scripts]
autoreadme = "autoreadme.cli:cli"

[project.urls]
Homepage = "https://autoreadme.dev"
Repository = "https://github.com/autoreadme/autoreadme"
Documentation = "https://docs.autoreadme.dev"
"Bug Tracker" = "https://github.com/autoreadme/autoreadme/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["autoreadme*"]

[tool.setuptools.package-data]
autoreadme = ["templates/*.md"]
